# Ensure we close pending tchannel requests.