--- a/benchmarks/benchserver/main.go
+++ b/benchmarks/benchserver/main.go
@@ -90,7 +90,7 @@ func serverTChannel(logger *zap.Logger) *testBackend.TestTChannelBackend {
 	simpleServiceCallHandler := baz.NewSimpleServiceCallHandler(handleSimpleServiceCall)
 
 	// must register handler first before bootstrap
-	tchannelBackend.Register("SimpleService", "Call", simpleServiceCallHandler)
+	tchannelBackend.Register("SimpleService", "call", simpleServiceCallHandler)
 
 	err = tchannelBackend.Bootstrap()
 	if err != nil {