--- a/runtime/tchannel_client.go
+++ b/runtime/tchannel_client.go
@@ -59,7 +59,9 @@ func NewTChannelClient(ch *tchannel.Channel, opt *TChannelClientOption) TChannel
 }
 
 func (c *tchannelClient) writeArgs(call *tchannel.OutboundCall, headers map[string]string, req RWTStruct) error {
-	writer, err := call.Arg2Writer()
+	twriter, err := call.Arg2Writer()
+	writer := DoubleCloseWriter{WriteFlusher: twriter}
+	defer writer.HardClose()
 	if err != nil {
 		return errors.Wrapf(err, "could not create arg2writer for outbound call %s: ", c.serviceName)
 	}
@@ -71,7 +73,9 @@ func (c *tchannelClient) writeArgs(call *tchannel.OutboundCall, headers map[stri
 		return errors.Wrapf(err, "could not close arg2writer for outbound call %s: ", c.serviceName)
 	}
 
-	writer, err = call.Arg3Writer()
+	twriter, err = call.Arg3Writer()
+	writer = DoubleCloseWriter{WriteFlusher: twriter}
+	defer writer.HardClose()
 	if err != nil {
 		return errors.Wrapf(err, "could not create arg3writer for outbound call %s: ", c.serviceName)
 	}
@@ -86,7 +90,9 @@ func (c *tchannelClient) writeArgs(call *tchannel.OutboundCall, headers map[stri
 // readResponse reads the response struct into resp, and returns:
 // (response headers, whether there was an application error, unexpected error).
 func (c *tchannelClient) readResponse(response *tchannel.OutboundCallResponse, resp RWTStruct) (bool, map[string]string, error) {
-	reader, err := response.Arg2Reader()
+	treader, err := response.Arg2Reader()
+	reader := DoubleCloseReader{ReadCloser: treader}
+	defer reader.HardClose()
 	if err != nil {
 		// Do not wrap system errors.
 		if _, ok := err.(tchannel.SystemError); ok {
@@ -110,7 +116,9 @@ func (c *tchannelClient) readResponse(response *tchannel.OutboundCallResponse, r
 	}
 
 	success := !response.ApplicationError()
-	reader, err = response.Arg3Reader()
+	treader, err = response.Arg3Reader()
+	reader = DoubleCloseReader{ReadCloser: treader}
+	defer reader.HardClose()
 	if err != nil {
 		return success, headers, errors.Wrapf(err, "could not create arg3Reader for outbound call response: %s", c.serviceName)
 	}