… runtime/tchannel_logger_test.go
… test/endpoints/baz/baz_simpleservice_method_compare_test.go
… test/endpoints/baz/baz_simpleservice_method_ping_test.go
… test/lib/test_backend/test_tchannel_backend.go
… test/lib/test_gateway/test_gateway_process.go