- For the server, make sure we close reader/writer.
  - For the client, make sure we close reader/writer.
  - For client/server, do thrift validation into a buffer
        and do not write an invalid partial RPC request down the wire

r: @uber/zanzibar-team