@Raynos 2017-07-11
@ChuntaoLu will write benchmarks.

as for `defer` vs manually calling `Close()` in every error case, i suspect that using `defer` is more go idiomatic. If we add a manual `Close()` in every error case then someone might add a new error case and forgot to close it.