@Raynos 2017-07-12
```
~/gocode/src/github.com/uber/zanzibar on tchannel-always-close*
$ benchcmp -best old.txt new.txt 
benchmark                              old ns/op     new ns/op     delta
BenchmarkHealthCall-2                  51845         51987         +0.27%
BenchmarkCall-2                        187711        202918        +8.10%
BenchmarkCompare-2                     172253        190383        +10.53%
BenchmarkPing-2                        157126        167176        +6.40%
BenchmarkSillyNoop-2                   148573        162854        +9.61%
BenchmarkSaveContacts-2                118151        117298        -0.72%
BenchmarkGoogleNowAddCredentials-2     97425         98834         +1.45%

benchmark                              old allocs     new allocs     delta
BenchmarkHealthCall-2                  125            125            +0.00%
BenchmarkCall-2                        364            382            +4.95%
BenchmarkCompare-2                     367            385            +4.90%
BenchmarkPing-2                        320            338            +5.62%
BenchmarkSillyNoop-2                   306            324            +5.88%
BenchmarkSaveContacts-2                331            331            +0.00%
BenchmarkGoogleNowAddCredentials-2     247            247            +0.00%

benchmark                              old bytes     new bytes     delta
BenchmarkHealthCall-2                  11098         11098         +0.00%
BenchmarkCall-2                        31450         33512         +6.56%
BenchmarkCompare-2                     31557         33612         +6.51%
BenchmarkPing-2                        27122         29180         +7.59%
BenchmarkSillyNoop-2                   26331         28387         +7.81%
BenchmarkSaveContacts-2                28336         28336         +0.00%
BenchmarkGoogleNowAddCredentials-2     20610         20611         +0.00%
```

It looks like writing into a buffer has a 6-9% performance regression in CPU. Will run other benchmark to look at that too.