@Raynos 2017-07-12
I verified the compare_to.sh script

```
uber at smart-rapide-9  
~/gocode/src/github.com/uber/zanzibar on tchannel-always-close*
$ benchcmp -best bench_tchannel-always-close.txt bench_master.txt 
benchmark                              old ns/op     new ns/op     delta
BenchmarkHealthCall-2                  53010         52287         -1.36%
BenchmarkCall-2                        202786        193423        -4.62%
BenchmarkCompare-2                     186833        174357        -6.68%
BenchmarkPing-2                        164019        158327        -3.47%
BenchmarkSillyNoop-2                   154382        150668        -2.41%
BenchmarkSaveContacts-2                118349        118229        -0.10%
BenchmarkGoogleNowAddCredentials-2     98242         97858         -0.39%

benchmark                              old allocs     new allocs     delta
BenchmarkHealthCall-2                  125            125            +0.00%
BenchmarkCall-2                        372            364            -2.15%
BenchmarkCompare-2                     375            367            -2.13%
BenchmarkPing-2                        328            320            -2.44%
BenchmarkSillyNoop-2                   314            306            -2.55%
BenchmarkSaveContacts-2                331            331            +0.00%
BenchmarkGoogleNowAddCredentials-2     247            247            +0.00%

benchmark                              old bytes     new bytes     delta
BenchmarkHealthCall-2                  11098         11097         -0.01%
BenchmarkCall-2                        33199         31460         -5.24%
BenchmarkCompare-2                     33295         31553         -5.23%
BenchmarkPing-2                        28862         27123         -6.03%
BenchmarkSillyNoop-2                   28067         26425         -5.85%
BenchmarkSaveContacts-2                28336         28332         -0.01%
BenchmarkGoogleNowAddCredentials-2     20610         20609         -0.00%
```

Master is about 3-6% faster.